Refactored controller layout. Added User and Admin controller along with their respective functions. Moved the views to their new folders and changed buttons to route to new changes.
