+++ petStoreMonitoringApp/Views/Home/Index.cshtml
@@ -17,8 +17,8 @@
         </div>
         <div class="container">
             <h2>Go to actions</h2>
-            <a href="Home/User" class="btn btn-primary btn-lg active btn-lg btn-block" role="button" aria-pressed="true">User</a>
-            <a href="Home/Admin" class="btn btn-secondary btn-lg active btn-lg btn-block" role="button" aria-pressed="true">Admin</a>
+            <a href="User/Index" class="btn btn-primary btn-lg active btn-lg btn-block" role="button" aria-pressed="true">User</a>
+            <a href="Admin/Index" class="btn btn-secondary btn-lg active btn-lg btn-block" role="button" aria-pressed="true">Admin</a>
         </div>
     </body>
 </html>
